chnages

+++ src/components/Header.jsx
@@ -1,16 +1,49 @@
 import React from 'react'
-import { Link, useLocation } from 'react-router-dom'
+import { Link, useLocation ,useNavigate } from 'react-router-dom'
 
 function Header() {
   const location = useLocation()
 
   const authData = JSON.parse(localStorage.getItem("authData"));
   const outletName = authData?.outlet_name; // Get 
-  const handleLogout = () => {
-    // Clear all localStorage items
-    localStorage.clear();
-    // The Link component will handle the navigation to /login
+  const userId = authData?.user_id; // Get 
+  const navigate = useNavigate();
+  const handleLogout = async () => {
+    try {
+
+    
+  
+      const logoutData = {
+        user_id: userId,
+        role: "cds",
+        app: "cds", // Assuming 'owner' is the application name
+      };
+  
+      // Make API request to logout using fetch
+      const response = await fetch("https://men4u.xyz/common_api/logout", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(logoutData), // Send logout data as JSON in the body
+      });
+  
+      const data = await response.json(); // Parse JSON response
+  
+      if (data.st === 1) {
+        // Successfully logged out, clear all localStorage items
+        localStorage.clear();
+        // Redirect to login page
+       navigate("/login")
+      } else {
+      }
+    } catch (error) {
+      console.error("Error logging out:", error);
+      window.showToast("error", error.message || "Failed to log out.");
+    }
   };
+  
+  
   return (
     <header className="bg-dark">
       <nav className="navbar navbar-expand-lg navbar-dark py-2">
@@ -23,9 +56,9 @@ function Header() {
           {/* Navigation Links */}
           <ul className="navbar-nav ms-auto align-items-center">
             <li className="nav-item fs-3">
-              <Link to="/login" className={`nav-link px-3`} onClick={handleLogout}>
+              <div className={`nav-link px-3`} onClick={handleLogout}>
                 <i className="bx bx-log-out"></i>
-              </Link>
+              </div>
             </li>
           </ul>
         </div>
